Implement organization mock in query
